requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
crispy-tailwind==1.0.3
cryptography==43.0.1
daphne==4.1.2
diff-match-patch==20230430
Django==5.1.1
django-allauth==64.2.1
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-guardian==2.4.0
django-import-export==4.1.1
django-phonenumber-field==8.0.0
django-tinymce==4.1.0
django-unfold==0.38.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djhtml==3.0.6
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isort==5.13.2
Markdown==3.7
phonenumbers==8.13.45
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
requests==2.32.3
service-identity==24.1.0
setuptools==75.0.0
sqlparse==0.5.1
tablib==3.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
zope.interface==7.0.3
